Add frontend files
The frontend has been setup with it's own package file to manage dependencies separately from the backend api app. Right now, it is setup in a way that you will have to run the compiling separately and a page reload will pickup the new app. All dependencies are compiled into a single bundled file for production (probably needs to be minified/uglified in the long run).